+++ extensions/google-chrome-profiles/src/util/clipboard.ts
@@ -1,7 +1,0 @@
-import { runAppleScript } from "run-applescript";
-
-export const getClipboardText = async () => {
-  const clipboard = await runAppleScript("the clipboard");
-  if (clipboard.length === 0) throw "Clipboard is empty";
-  else return clipboard;
-};