+++ extensions/spotify-player/src/spotify/interfaces.ts
@@ -1,0 +1,12 @@
+export interface Response<T> {
+  result?: T;
+  error?: string;
+  isLoading?: boolean;
+}
+
+export interface AuthResponseCredentials {
+  access_token: string;
+  token_type: string;
+  expires_in: number;
+  scope: string;
+}